requirements.txt
setuptools
wheel
customtkinter
yt_dlp
tkcalendar
requests
gTTS
translate
python-docx
odfpy
openpyxl
PyGithub
pyperclip
websocket-client
pillow
playsound3
speechRecognition
meteofrance_api
pyradios
python-vlc
pyspellchecker
llama_cpp_python
pyttsx3